requirements.txt
# Core dependencies
numpy>=1.20.0
opencv-python>=4.5.5
Pillow>=9.0.0
matplotlib>=3.5.0
tqdm>=4.62.0

# Deep learning frameworks
torch>=1.10.0
torchvision>=0.11.0
timm>=0.6.7  # Transformer models

# Data handling
pandas>=1.3.0
requests>=2.27.0

# UI components
customtkinter>=4.6.0

# Image processing
scikit-image>=0.19.0

# Utility
pathlib>=1.0.1

# Optional - for YOLO detection
ultralytics>=8.0.0  # Uncomment if you want to use YOLO detection

